requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
passlib[bcrypt]